Set dimensions for scaler variables to --none--
Some variables in CfRadial and NCAS-RADAR require variables to either be scaler (i.e. have no dimensions) or have a defined dimension,
which may result in the need to check the dimensions of a scaler variable. Previously, with scaler variables having dimensions listed
as "" in the output from CDLReader.to_dict() function, trying to check dimensions would say it is undefined. This change allows checking
whether a variable is a scaler (defined_attrs dimension:--none--) and allows checking if a variable is a scaler or has a defined dimension
(rules_attrs dimension: rule-func:match-one-of:--none--|time).